// packages/shared/src/workflow/types.ts
export interface ExecutionContext {
  botId: string;
  commandName: string;
  user: {
    id: string;
    username: string;
    discriminator?: string;
  };
  guild?: {
    id: string;
    name?: string;
    iconUrl?: string;
  };
  channel?: {
    id: string;
    name?: string;
  };
  options: Record<string, unknown>;
  memberRoles: string[];
  variables: Record<string, unknown>;
}